Streamlined for persistence
'Streamlined' the program structure for better persistence.
Primarily, this involved changing dynamically registered broadcast receivers to manifest-specified receivers that spawn the main service.
Also, the alarmmanager's inexact repeating alarm for persistence was set to the 'exact' type depending on the Android version and the respawn interval was changed to 1 minute. And other ancillary changes.